refactor(render): introduce RenderSession — consolidate widget passes (#1145)
Extracts per-frame rendering state from ProcessApp into a dedicated
RenderSession struct. All three render passes (painter commands,
TextInput widgets, scroll-region scan) now execute inside
RenderSession::render(), with accumulated events drained back to
ProcessApp after the frame. ProcessApp loses five direct fields
(text_input_buffers, text_input_visible_prev, text_input_has_focus,
pane_just_focused, scroll_offsets) — all now owned by render_session.
